Update Unit 3 and add Unit 4

+++ Unit3/src/main/java/com/javaunit3/springmvc/model/MovieEntity.java
@@ -1,15 +1,28 @@
 package com.javaunit3.springmvc.model;
 
-import javax.persistence.*;
+
 import java.util.List;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+
+
 
 @Entity
 @Table(name = "movies")
 public class MovieEntity {
+	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "movie_Id")
-	private Integer id;
+	private Long id;
 
 	@Column(name = "title")
 	private String title;
@@ -36,11 +49,11 @@ public class MovieEntity {
 		this.votes.add(vote);
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
